Fixes warning:
ReadGDSIIFile.cc:622:19: warning: cast from 'libGDSII::BYTE *' (aka 'unsigned char *') to
      'libGDSII::WORD *' (aka 'unsigned short *') increases required alignment from 1 to 2
      [-Wcast-align]
        WORD W = *(WORD *)Payload;

Rather than perform a pointer cast, just explicitly load two bytes and shift them
to make a 16-bit word.   This has the additional advantage of being more portable,
because we can specify the endian-ness with which we read the data.

I used a little-endian order for backward compatibility with x86 machines.
(It's not clear to me whether the GDSII format specifies a particular
interpretation of bitfield indices, so that endianness matters here?)